requirements.txt
# Testing
pytest
pytest-aiohttp
pytest-asyncio
